+++ CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to this project will be documented in this file.
 * Add spectral analysis module.
 * Add state variable filter.
 * Introduce signal abstraction and basic operations on it.
+* Implement sine and static signal generators.